ecommerce_beckend/requirements.txt
Django==5.0.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.3
drf-spectacular==0.26.5
django-celery-beat==2.5.0
django-celery-results==2.5.1
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
django-environ==0.11.2
Pillow==10.1.0
gunicorn==21.2.0
whitenoise==6.6.0
django-ratelimit==4.1.0
sentry-sdk==1.38.0
